セグフォしたのでunsinged longに戻した

+++ matmul+cg/host/src/main_gpu.cu
@@ -5,10 +5,10 @@
 #include <chrono>
 // #include "calc_on_fpga.h"
 
-__global__ void matmul(float *a, float *b, float *c, unsigned int n) {
-  unsigned int j = blockIdx.x * blockDim.x + threadIdx.x; // 通し番号を得るための計算
-  unsigned int i = blockIdx.y * blockDim.y + threadIdx.y;
-  unsigned int k;
+__global__ void matmul(float *a, float *b, float *c, unsigned long n) {
+  unsigned long j = blockIdx.x * blockDim.x + threadIdx.x; // 通し番号を得るための計算
+  unsigned long i = blockIdx.y * blockDim.y + threadIdx.y;
+  unsigned long k;
   float sum = 0.0f;
   if (i < n || j < n) {
     for(k=0; k<n; k++) {
@@ -25,10 +25,10 @@ int main(int argc, char *argv[]) {
   if (argc != 5) { std::cerr << "Error! The number of arguments is wrong."              << std::endl; exit(1); }
 
   const char *name     = argv[1];
-  const unsigned int numdata_h = std::stoull(std::string(argv[2]));
+  const unsigned long numdata_h = std::stoull(std::string(argv[2]));
   const int  numstream = std::stoull(std::string(argv[3]));
   const int  numtry    = std::stoull(std::string(argv[4]));
-  const unsigned int numbyte   = numdata_h * numdata_h * sizeof(float); // this sample uses "float"
+  const unsigned long numbyte   = numdata_h * numdata_h * sizeof(float); // this sample uses "float"
 
   // size_t global_item_size[3];
   // size_t local_item_size[3];
@@ -38,15 +38,15 @@ int main(int argc, char *argv[]) {
 
   /***** GPU *****/
   static const int numthread = 256;  
-  const unsigned int numblock = (numdata_h % numthread) ? (numdata_h/numthread) + 1 : (numdata_h/numthread);
+  const unsigned long numblock = (numdata_h % numthread) ? (numdata_h/numthread) + 1 : (numdata_h/numthread);
   float *h_a, *h_b, *h_c;
 
   cudaMallocHost(&h_a, numbyte);
   cudaMallocHost(&h_b, numbyte);
   cudaMallocHost(&h_c, numbyte);
   
-  for (unsigned int i = 0; i < numdata_h; i++) {
-    for (unsigned int j = 0; j < numdata_h; j++) {
+  for (unsigned long i = 0; i < numdata_h; i++) {
+    for (unsigned long j = 0; j < numdata_h; j++) {
       h_a[i*numdata_h+j] = 0.0f; //(j+1)/2*0.0001f;
       h_b[i*numdata_h+j] = 0.5f;
       h_c[i*numdata_h+j] = 0.0f;
